+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 This file is used to list changes made in each version of the kapacitor cookbook.
 
+0.0.7
+-----
+- Cyberflow - add ability to update handlers
+
+0.0.6
+-----
+- Cyberflow - fix hendler api endpoints for 1.4.0.kapacitor version
+
+0.0.5
+-----
+- Cyberflow - add ability to rewrite kapacitor tasks
+
+0.0.3
+-----
+- Cyberflow - add kapacitor_handler resource
+
 0.0.1
 -----